+++ enjoy/src/constants/gpt-presets.ts
@@ -1,137 +1,4 @@
-export const DATABASE_NAME = "enjoy_database";
-export const LIBRARY_PATH_SUFFIX = "EnjoyLibrary";
-
-export const STORAGE_WORKER_ENDPOINT = "https://storage.enjoy.bot";
-export const STORAGE_WORKER_ENDPOINTS = [
-  "https://storage.enjoy.bot",
-  "https://enjoy-storage.baizhiheizi.com",
-];
-
-export const AI_WORKER_ENDPOINT = "https://ai-worker.enjoy.bot";
-
-export const WEB_API_URL = "https://enjoy.bot";
-export const WEB_API_URLS = ["https://enjoy.bot", "https://enjoy-web.fly.dev"];
-
-export const REPO_URL = "https://github.com/xiaolai/everyone-can-use-english";
-
-export const SENTRY_DSN =
-  "https://d51056d7af7d14eae446c0c15b4f3d31@o1168905.ingest.us.sentry.io/4506969353289728";
-
-export const MAGIC_TOKEN_REGEX =
-  /\b(Mrs|Ms|Mr|Dr|Prof|St|[a-zA-Z]{1,2}|\d{1,2})\.\b/g;
-export const END_OF_SENTENCE_REGEX = /[^\.!,\?][\.!\?]/g;
-
-export const FFMPEG_TRIM_SILENCE_OPTIONS = [
-  "-af",
-  "silenceremove=1:start_duration=1:start_threshold=-50dB:detection=peak,aformat=dblp,areverse,silenceremove=start_periods=1:start_duration=1:start_threshold=-50dB:detection=peak,aformat=dblp,areverse",
-];
-
-export const FFMPEG_CONVERT_WAV_OPTIONS = [
-  "-ar",
-  "16000",
-  "-ac",
-  "1",
-  "-c:a",
-  "pcm_s16le",
-];
-
-// https://hf-mirror.com/ggerganov/whisper.cpp/tree/main
-export const WHISPER_MODELS_OPTIONS = [
-  {
-    type: "tiny",
-    name: "ggml-tiny.bin",
-    size: "75 MB",
-    sha: "bd577a113a864445d4c299885e0cb97d4ba92b5f",
-    url: "https://hf-mirror.com/ggerganov/whisper.cpp/resolve/main/ggml-tiny.bin",
-  },
-  {
-    type: "tiny.en",
-    name: "ggml-tiny.en.bin",
-    size: "75 MB",
-    sha: "c78c86eb1a8faa21b369bcd33207cc90d64ae9df",
-    url: "https://hf-mirror.com/ggerganov/whisper.cpp/resolve/main/ggml-tiny.en.bin",
-  },
-  {
-    type: "base",
-    name: "ggml-base.bin",
-    size: "142 MB",
-    sha: "465707469ff3a37a2b9b8d8f89f2f99de7299dac",
-    url: "https://hf-mirror.com/ggerganov/whisper.cpp/resolve/main/ggml-base.bin",
-  },
-  {
-    type: "base.en",
-    name: "ggml-base.en.bin",
-    size: "142 MB",
-    sha: "137c40403d78fd54d454da0f9bd998f78703390c",
-    url: "https://hf-mirror.com/ggerganov/whisper.cpp/resolve/main/ggml-base.en.bin",
-  },
-  {
-    type: "small",
-    name: "ggml-small.bin",
-    size: "466 MB",
-    sha: "55356645c2b361a969dfd0ef2c5a50d530afd8d5",
-    url: "https://hf-mirror.com/ggerganov/whisper.cpp/resolve/main/ggml-small.bin",
-  },
-  {
-    type: "small.en",
-    name: "ggml-small.en.bin",
-    size: "466 MB",
-    sha: "db8a495a91d927739e50b3fc1cc4c6b8f6c2d022",
-    url: "https://hf-mirror.com/ggerganov/whisper.cpp/resolve/main/ggml-small.en.bin",
-  },
-  {
-    type: "medium",
-    name: "ggml-medium.bin",
-    size: "1.5 GB",
-    sha: "fd9727b6e1217c2f614f9b698455c4ffd82463b4",
-    url: "https://hf-mirror.com/ggerganov/whisper.cpp/resolve/main/ggml-medium.bin",
-  },
-  {
-    type: "medium.en",
-    name: "ggml-medium.en.bin",
-    size: "1.5 GB",
-    sha: "8c30f0e44ce9560643ebd10bbe50cd20eafd3723",
-    url: "https://hf-mirror.com/ggerganov/whisper.cpp/resolve/main/ggml-medium.en.bin",
-  },
-  {
-    type: "large-v1",
-    name: "ggml-large-v1.bin",
-    size: "2.9 GB",
-    sha: "b1caaf735c4cc1429223d5a74f0f4d0b9b59a299",
-    url: "https://hf-mirror.com/ggerganov/whisper.cpp/resolve/main/ggml-large-v1.bin",
-  },
-  {
-    type: "large-v2",
-    name: "ggml-large-v2.bin",
-    size: "2.9 GB",
-    sha: "0f4c8e34f21cf1a914c59d8b3ce882345ad349d6",
-    url: "https://hf-mirror.com/ggerganov/whisper.cpp/resolve/main/ggml-large-v2.bin",
-  },
-  {
-    type: "large",
-    name: "ggml-large-v3.bin",
-    size: "2.9 GB",
-    sha: "ad82bf6a9043ceed055076d0fd39f5f186ff8062",
-    url: "https://hf-mirror.com/ggerganov/whisper.cpp/resolve/main/ggml-large-v3.bin",
-  },
-];
-
-export const AudioFormats = ["mp3", "wav", "ogg", "flac", "m4a", "wma", "aac"];
-
-export const VideoFormats = ["mp4", "mkv", "avi", "mov", "wmv", "flv", "webm"];
-
-export const PROCESS_TIMEOUT = 1000 * 60 * 15;
-
-export const AI_GATEWAY_ENDPOINT =
-  "https://gateway.ai.cloudflare.com/v1/11d43ab275eb7e1b271ba4089ecc3864/enjoy";
-
-export const NOT_SUPPORT_JSON_FORMAT_MODELS = [
-  "gpt-4-vision-preview",
-  "gpt-4",
-  "gpt-4-32k",
-];
-
-export const CONVERSATION_PRESETS = [
+export const GPT_PRESETS = [
   {
     key: "english-coach",
     name: "英语教练",
@@ -369,219 +236,3 @@ export const CONVERSATION_PRESETS = [
     },
   },
 ];
-
-export const IPA_CONSONANTS: { [key: string]: string[] } = {
-  plosive: [
-    "p",
-    "b",
-    "t",
-    "d",
-    "ʈ",
-    "ɖ",
-    "c",
-    "ɟ",
-    "k",
-    "g",
-    "q",
-    "ɢ",
-    "ʔ",
-    /* extensions */ "ɡ",
-  ],
-  nasal: ["m", "ɱ", "n", "ɳ", "ɲ", "ŋ", "ɴ", "n̩"],
-  trill: ["ʙ", "r", "ʀ"],
-  tapOrFlap: ["ⱱ", "ɾ", "ɽ"],
-  fricative: [
-    "ɸ",
-    "β",
-    "f",
-    "v",
-    "θ",
-    "ð",
-    "s",
-    "z",
-    "ʃ",
-    "ʒ",
-    "ʂ",
-    "ʐ",
-    "ç",
-    "ʝ",
-    "x",
-    "ɣ",
-    "χ",
-    "ʁ",
-    "ħ",
-    "ʕ",
-    "h",
-    "ɦ",
-  ],
-  lateralFricative: ["ɬ", "ɮ"],
-  affricate: ["tʃ", "ʈʃ", "dʒ"], // very incomplete, there are many others
-  approximant: ["ʋ", "ɹ", "ɻ", "j", "ɰ", /* extensions */ "w"],
-  lateralApproximant: ["l", "ɭ", "ʎ", "ʟ"],
-};
-
-export const IPA_VOWELS: { [key: string]: string[] } = {
-  close: ["i", "yɨ", "ʉɯ", "u", "iː"],
-  closeOther: ["ɪ", "ʏ", "ʊ", "ɨ", "ᵻ"],
-  closeMid: ["e", "ø", "ɘ", "ɵ", "ɤ", "o", "ə", "oː"],
-  openMid: ["ɛ", "œ", "ɜ", "ɞ", "ʌ", "ɔ", "ɜː", "uː", "ɔː", "ɛː"],
-  open: ["æ", "a", "ɶ", "ɐ", "ɑ", "ɒ", "ɑː"],
-  rhotic: ["◌˞", "ɚ", "ɝ", "ɹ̩"],
-  diphtongs: [
-    "eɪ",
-    "əʊ",
-    "oʊ",
-    "aɪ",
-    "ɔɪ",
-    "aʊ",
-    "iə",
-    "ɜr",
-    "ɑr",
-    "ɔr",
-    "oʊr",
-    "oːɹ",
-    "ir",
-    "ɪɹ",
-    "ɔːɹ",
-    "ɑːɹ",
-    "ʊɹ",
-    "ʊr",
-    "ɛr",
-    "ɛɹ",
-    "əl",
-    "aɪɚ",
-    "aɪə",
-  ],
-};
-
-export const IPA_MAPPINGS: { [key: string]: string } = {
-  p: "p",
-  b: "b",
-  t: "t",
-  d: "d",
-  ʈ: "t",
-  ɖ: "d",
-  c: "k",
-  ɟ: "g",
-  k: "k",
-  g: "g",
-  q: "k",
-  ɢ: "g",
-  ʔ: "t",
-  ɡ: "g",
-  m: "m",
-  ɱ: "m",
-  n: "n",
-  ɳ: "n",
-  ɲ: "j",
-  ŋ: "ŋ",
-  ɴ: "ŋ",
-  n̩: "n",
-  ʙ: "r",
-  r: "r",
-  ʀ: "r",
-  ⱱ: "",
-  ɾ: "t",
-  ɽ: "r",
-  ɸ: "f",
-  β: "v",
-  f: "f",
-  v: "v",
-  θ: "θ",
-  ð: "ð",
-  s: "s",
-  z: "z",
-  ʃ: "ʃ",
-  ʒ: "ʒ",
-  ʂ: "s",
-  ʐ: "z",
-  ç: "",
-  ʝ: "j",
-  x: "k",
-  ɣ: "g",
-  χ: "h",
-  ʁ: "r",
-  ħ: "h",
-  ʕ: "",
-  h: "h",
-  ɦ: "h",
-  ɬ: "",
-  ɮ: "",
-  tʃ: "tʃ",
-  ʈʃ: "tʃ",
-  dʒ: "dʒ",
-  ʋ: "v",
-  ɹ: "r",
-  ɻ: "r",
-  j: "j",
-  ɰ: "w",
-  w: "w",
-  l: "l",
-  ɭ: "l",
-  ʎ: "j",
-  ʟ: "l",
-  i: "i",
-  yɨ: "iː",
-  ʉɯ: "uː",
-  u: "uː",
-  iː: "iː",
-  ɪ: "ɪ",
-  ʏ: "ɪ",
-  ʊ: "ʊ",
-  ɨ: "i",
-  ᵻ: "i:",
-  e: "e",
-  ø: "e",
-  ɘ: "ə",
-  ɵ: "ə",
-  ɤ: "ɑː",
-  o: "o",
-  ə: "ə",
-  oː: "oː",
-  ɛ: "e",
-  œ: "æ",
-  ɜ: "ɝ",
-  ɞ: "əː",
-  ʌ: "ʌ",
-  ɔ: "ɔ",
-  ɜː: "ɝː",
-  uː: "uː",
-  ɔː: "ɔː",
-  ɛː: "e:",
-  eː: "i:",
-  æ: "æ",
-  a: "ɑ",
-  ɶ: "ɑ",
-  ɐ: "ə",
-  ɑ: "ɑ",
-  ɒ: "ɑː",
-  ɑː: "ɑː",
-  "◌˞": "",
-  ɚ: "ɚ",
-  ɝ: "ɝ",
-  ɹ̩: "r",
-  eɪ: "eɪ",
-  əʊ: "oʊ",
-  oʊ: "oʊ",
-  aɪ: "aɪ",
-  ɔɪ: "ɔɪ",
-  aʊ: "aʊ",
-  iə: "iə",
-  ɜr: "ɜr",
-  ɑr: "ɑr",
-  ɔr: "ɔr",
-  oʊr: "oʊr",
-  oːɹ: "ɔːr",
-  ir: "ir",
-  ɪɹ: "ɪr",
-  ɔːɹ: "ɔːr",
-  ɑːɹ: "ɑːr",
-  ʊɹ: "ʊr",
-  ʊr: "ʊr",
-  ɛr: "er",
-  ɛɹ: "er",
-  əl: "əl",
-  aɪɚ: "aɪ",
-  aɪə: "aɪə",
-  ts: "tz",
-};
